# Make database operations 30x faster by using SQLite