--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 test.log
 error.log
 rebalance_db.txt
+rebalance.db
 testing_data