## Preface
Rebalancing is a lengthy process and thus ensuring it can be resumed is rather important. Currently, the code implements
a simple database-like textual file. When script is interrupted and then subsequently restarted, the rebalancing process
can be restarted more-or-less from where it was interrupted.


## The problem
Current solution is great conceptually. However, the implementation suffers from very poor performance due to multiple
factors:

 - spawning multiple processes carries a non-insignificant cost
 - re-reading a good chunk of the database file generates a lot of I/O, even if in-memory
 - the data structure, due ot its simplicity and lack of indexing does not scale well with large number of files

The result of this is a very poor performance during both rebalance and when script needs to be resumed. In the case of
my dataset with ~1M files resume from 500k point took close to 12 hours(!). The performance problems are magnified with
small files being present


## Solutions evaluation
The first optimization I attempted was limiting re-reads for counting. However, this still proven very limiting. The
next idea I tested was using `bash` associative arrays, implementing an in-memory cache persisted occasionally to the
disk. Unfortunately, with 500k+ files the memory usage started to creep into 500-600-800MB range, with more and more
being ostensibly leaked. This shows clearly that bash array were never designed for such a dataset.

I scraped the idea and came back to the drawing board. After ad-hoc testing, I implemented a prototype based on SQLite
database, that is stored in a dedicated file as well. The time for read-save operations were cut by **orders of magnitude**,
despite still launching at least 2 processes for every file. As a next step I implemented a delayed database flush, with
60s timeout, further cutting the db operations time.


## Potential impact for users
The change carries some impact to systems utilizing this project.

 - SQLite database is a binary blob, that is ~30-40% larger than a text file
   - while the database is no longer readable as a text file, the `sqlite3` tool offers an easy interface
   - `README.md` contains a simulated transaction, so that unfamiliar users can easily interact with the db 
 - The usage of SQLite introduces a new dependency on `sqlite3`
   - `sqlite3` is already present on systems that are most likely the target of this script
   - I verified existence of a compatible `sqlite3` binary on TrueNAS SCALE, TrueNAS CORE, macOS 13, as well as Ubuntu
   - Debian doesn't ship with `sqlite3` by default, but package is in the default repo
 - With the delayed-flush of the database to disk, there's a potential of rebalance state being lost
   - at most 60 seconds of the progress can be lost when interrupted
   - re-balancing a file more times than requested doesn't carry a risk, thus I see this as non-issue
 - Previously built databases are no longer usable
   - this PR includes `convert-legacy-db.sh` that handles the migration
   - the script when started in presence of the old database will abort to ensure balancing doesn't start from scratch
     unintentionally
   - documentations has been updated to cover that


## Testing methodology
  - I used my preexisting database file where most of the files will be hit
    - database starting point is being equal for all scripts
    - new version gets `cp rebalance_db.txt_copy rebalance_db.txt ; ./convert-legacy-db.sh` (this takes ~35 seconds)
    - old version gets `cp rebalance_db.txt_copy rebalance_db.txt`
    - in other words, both versions operate on the same logical set of data
  - benchmark pure database operations (commented-out cp/mv/rm)
  - the pool uses HDDs with SSDs for metadata and contains ~500k files
  - command run: `time ./zfs-inplace-rebalancing.sh --checksum false --passes 2 /mnt/testdata > rebalance.log`


## Test results
 - current `.txt`-based database: **29h 14min**
 - SQLite database with per-file flush: **1h 32min**
 - SQLite database with delayed flush every ~60 seconds: **56min**

In other words, it's almost a 30x speed-up :)
There are a few smaller low-hanging fruits for optimization, but I'm tackling one thing at a time.


## WDYT?
...about this PR and the idea itself? Marking as `draft` as of right now as I didn't have a chance to test on FreeBSD
yet.