@markusressel 2023-09-09
Hey @kiler129 ! :nerd_face: 
Thx for the really in-depth analysis and work you put into this :astonished: :heart: 

I have been thinking about this for a bit and it really looks like a great improvement to the script. I have two thoughts we may want to discuss:

## 1. Should we keep both the old/current as well as the new script?

Since this script is a pretty big change to "how things work", as well as what a potential user has to know about that to be able to work with it (reading txt file vs sqlite db, etc.), I am wondering if we should keep the old script available for people with a small usecase, and add this version for bigger usecases. I think maintaining both versions shouldn't be that big of a deal (at least if you can keep an eye on it as well :sweat_smile: ).

## 2. Should we consider "moving" to a "real" programming language to implement this script?

Looking at how this version works, I am wondering if a shift to a "real" programming language that works on most systems, like f.ex. golang (for performance) or python (for convenience and accessibility) would be the "better" path for future optimization? The single reason I initially used Bash to implement this script is because it was a one-time operation for me, which I will probably never do again. Bash is hard though, especially optimizing, debugging and testing it. Utilizing a programming language, accessing a database, as well as providing the CLI tools to work with it would be much easier.

Note: Following this route does not necessitate "trashing" this PR, it might simply be a third version, or an evolution of this PR.