@kiler129 2023-09-10
> Hey @kiler129 ! 🤓 Thx for the really in-depth analysis and work you put into this 😲 ❤️

Don't we all start with "it will be just a small fix"? ...and that's just the beginning`**` 😄 



> ### 1. Should we keep both the old/current as well as the new script?

I was battling with this as well: should we make this a parameter... or maybe support both based on existence of the current "resume" file? What about the default? (should we feature-check or....)

This is exactly why I put this PR as a draft to see what's the best compromise for that. I think if the intention would be to keep the current database a smart way would be to:

1. **Default to SQLite so new users can benefit from the speed-up**
  - Feature-check for `sqlite3` and behave git-way if missing, i.e. tell the user that `sqlite3` isn't available and suggest that it should be installed. In addition provide an opt-in to force text file with a warning of the speed drawback, esp. problematic with large pools but I wouldn't suggest that in the error message as ideally you don't want people using `txt` if they can use `sqlite3` ;) 
  - The user should be able to decide to opt-in for textual database, even if `sqlite3` is available (but maybe it's broken/substandard/etc)
  - I believe this can be solved with an option `--db=[sqlite,txt]`. When option is present it will force the script to fail if e.g. `sqlite3` isn't available but `--db=sqlite` was passed.
2. **Handle old databases gracefully**
  - If old database is detected, put a clear/big warning "hey, it's suboptimal but I will continue with the old method". In essence this will behave like passing `--db=txt`, but will also suggest the db conversion or just running with `--db=sqlite`/renaming old db to start from scratch
  - If old db is present and `--db=sqlite` is specified, continue using `sqlite3` but also add a warning "hey, your old DB will be outdated, rename it to avoid confusion`

--


> ### 2. Should we consider "moving" to a "real" programming language to implement this script?

**The good: easy of maintenance**
My software engineer soul screams "yes!".... and I even started prototyping it from scratch. First I started with PHP due to its very vast SPL (as it should be a single file, not a dependency-hell) and good `select()` and process control support. Then I realized while our servers do have PHP, storage boxes usually don't and started considering other possibilities.

**The bad: portability**
....and then the realization was sad. Bash is horrible, but portable-ish. Perl would arguably be even a worse choice given the write-once read-never nature. While Python is available on most system, by design its standard library is very slim and `pip` isn't an option here. Additional complexity is the versions shipped, which in case of Python actually matters quite a bit, as I learned in my last project. In addition, Python is *painfully* slow with many non-obvious things (e.g. some mass-string operations that take 1-2s on slow PHP take over an HOUR in Python 3, while `select()` being performance-fragile).
However, I started digging deeper in a hope we can MAYBE at least justify this by not using `cp` or `mv` or other CLI tools with exec but actually use native functionality. However, it seems like nobody wants to deal with complexity of VFS et. al. in a general programing language: see the red box in [shutil.copy](https://docs.python.org/3/library/shutil.html#shutil.copy).

**The ugly: real world has no alternatives**
It seems like there are no other universal candidates sadly. I would love to use anything else: Go, Rust, Swift... you name it. But sadly it seems like Bash is the best compromise here, especially while working with appliances like TrueNAS that prevent users from installing any software. This is also why I maintain [this abomination of shell scripts](https://github.com/kiler129/server-healthchecks) for monitoring, even if doing that in a normal programming language would've been orders of magnitude easier.


---

**Off-topic:**
`**` Issue awaiting; I need to do much more testing but tl;dr the code works great if you can **ensure** nothing is writing to the filesystem. However, when you cannot be sure things get messy. Even with md5 check it's **not** safe to run it when processes can potentially write/have files open. At this point there are 3 software engineers/Linux admins poking at how to do it safely ;) If our ideas will work in practice on a test 20TB pool with many apps trying to grab files and we don't destroy any files, the costly checksum can probably be completely avoided + the code could **guarantee** data safety.